requirements.txt
segno==1.5.2
Pillow==10.0.1
pyzbar==0.1.9
colorama==0.4.6
